backend/requirements.txt
# FastAPI and server
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic==2.10.3
pydantic-settings==2.6.1

# Vector database and embeddings
lancedb==0.18.0
sentence-transformers==3.3.1
torch==2.5.1

# LLM integrations
anthropic==0.40.0
openai==1.56.0

# Document processing
PyMuPDF==1.25.1
python-docx==1.1.2
pdfplumber==0.11.4
python-pptx==1.0.2
openpyxl==3.1.5
xlrd==2.0.2
pandas==2.2.3

# Web scraping
trafilatura==1.12.2
requests==2.32.3
beautifulsoup4==4.12.3

# YouTube
youtube-transcript-api==0.6.3
moviepy==2.1.1

# Audio/Video transcription (requires ffmpeg: brew install ffmpeg)
faster-whisper==1.1.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
keyring==25.5.0
dateparser==1.2.0
